move now generic event checker to action_rules
