+++ acceptance_tests/features/steps/receipt.py
@@ -5,10 +5,7 @@ import time
 from behave import when, then, step
 from google.api_core.exceptions import GoogleAPIError
 from google.cloud import pubsub_v1
-from retrying import retry
 
-from acceptance_tests.features.steps.action_rules import setup_action_rule
-from acceptance_tests.features.steps.case_look_up import get_logged_events_for_case_by_id
 from acceptance_tests.utilities.rabbit_helper import start_listening_to_rabbit_queue, store_all_msgs_in_context
 from acceptance_tests.utilities.test_case_helper import test_helper
 from config import Config
@@ -165,20 +162,3 @@ def _publish_offline_receipt(context, tx_id="3d14675d-a25d-4672-a0fe-b960586653e
     print(f'Message published to {topic_path}')
 
     context.sent_to_gcp = True
-
-
-@step('set action rule of type "{action_type}" when case event "{event_type}" is logged')
-def set_action_rule_when_case_event_logged(context, action_type, event_type):
-    check_for_event(context, event_type)
-    setup_action_rule(context, action_type)
-
-
-@retry(stop_max_attempt_number=30, wait_fixed=1000)
-def check_for_event(context, event_type):
-    events = get_logged_events_for_case_by_id(context.case_created_events[0]['payload']['collectionCase']['id'])
-
-    for event in events:
-        if event['eventType'] == event_type:
-            return
-
-    test_helper.fail(f"Case {context.first_case['id']} event_type {event_type} not logged")
